// controlPanel/sdk/model/user.ts
/**
 * server
 * server
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    id?: number;
    name: string;
    email: string;
    creationDate: Date;
    lastUpdatedDate: Date;
    isBlocked: boolean;
    roles?: Array<string>;
}